Move files to dashboard package
