+++ pir/pir-impl/src/main/java/com/duckduckgo/pir/impl/dashboard/PirDashboardWebViewActivity.kt
@@ -35,13 +35,12 @@ import com.duckduckgo.js.messaging.api.JsMessageCallback
 import com.duckduckgo.js.messaging.api.JsMessaging
 import com.duckduckgo.js.messaging.api.SubscriptionEventData
 import com.duckduckgo.pir.api.dashboard.PirDashboardWebViewScreen
+import com.duckduckgo.pir.impl.dashboard.PirDashboardWebViewViewModel.Command
+import com.duckduckgo.pir.impl.dashboard.PirDashboardWebViewViewModel.Command.SendJsEvent
+import com.duckduckgo.pir.impl.dashboard.PirDashboardWebViewViewModel.Command.SendResponseToJs
+import com.duckduckgo.pir.impl.dashboard.messaging.PirDashboardWebConstants
+import com.duckduckgo.pir.impl.dashboard.messaging.PirDashboardWebViewClient
 import com.duckduckgo.pir.impl.databinding.ActivityPirDashboardWebviewBinding
-import com.duckduckgo.pir.impl.messaging.PirDashboardWebConstants
-import com.duckduckgo.pir.impl.ui.PirDashboardWebViewClient
-import com.duckduckgo.pir.impl.ui.PirDashboardWebViewViewModel
-import com.duckduckgo.pir.impl.ui.PirDashboardWebViewViewModel.Command
-import com.duckduckgo.pir.impl.ui.PirDashboardWebViewViewModel.Command.SendJsEvent
-import com.duckduckgo.pir.impl.ui.PirDashboardWebViewViewModel.Command.SendResponseToJs
 import javax.inject.Inject
 import javax.inject.Named
 import kotlinx.coroutines.flow.launchIn
