+++ pir/pir-impl/src/main/java/com/duckduckgo/pir/impl/dashboard/messaging/PirDashboardWebMessagingInterface.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.pir.impl.messaging
+package com.duckduckgo.pir.impl.dashboard.messaging
 
 import android.webkit.JavascriptInterface
 import android.webkit.WebView
@@ -32,7 +32,7 @@ import com.duckduckgo.js.messaging.api.JsRequestResponse
 import com.duckduckgo.js.messaging.api.SubscriptionEvent
 import com.duckduckgo.js.messaging.api.SubscriptionEventData
 import com.duckduckgo.pir.impl.brokers.JSONObjectAdapter
-import com.duckduckgo.pir.impl.messaging.handlers.PirWebJsMessageHandler
+import com.duckduckgo.pir.impl.dashboard.messaging.handlers.PirWebJsMessageHandler
 import com.squareup.anvil.annotations.ContributesBinding
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
