+++ pir/pir-impl/src/main/java/com/duckduckgo/pir/impl/dashboard/messaging/handlers/PirWebInitialScanStatusMessageHandler.kt
@@ -14,40 +14,40 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.pir.impl.messaging.handlers
+package com.duckduckgo.pir.impl.dashboard.messaging.handlers
 
 import com.duckduckgo.di.scopes.ActivityScope
 import com.duckduckgo.js.messaging.api.JsCallbackData
 import com.duckduckgo.js.messaging.api.JsMessage
 import com.duckduckgo.js.messaging.api.JsMessageCallback
 import com.duckduckgo.js.messaging.api.JsMessaging
-import com.duckduckgo.pir.impl.messaging.PirDashboardWebConstants
-import com.duckduckgo.pir.impl.messaging.PirDashboardWebMessages
+import com.duckduckgo.pir.impl.dashboard.messaging.PirDashboardWebConstants
+import com.duckduckgo.pir.impl.dashboard.messaging.PirDashboardWebMessages
 import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
 import logcat.logcat
+import org.json.JSONArray
 import org.json.JSONObject
 
 /**
- * Handles the initial getCurrentUserProfile message from Web which is used to retrieve the current user profile
- * and decide whether to show the onboarding or the dashboard.
+ * Handles the initial scan status message from Web which is used to retrieve the status of the initial scan.
  */
 @ContributesMultibinding(
     scope = ActivityScope::class,
     boundType = PirWebJsMessageHandler::class,
 )
-class PirWebGetCurrentUserProfileMessageHandler @Inject constructor() :
+class PirWebInitialScanStatusMessageHandler @Inject constructor() :
     PirWebJsMessageHandler() {
 
     override val methods: List<String> =
-        listOf(PirDashboardWebMessages.GET_CURRENT_USER_PROFILE.messageName)
+        listOf(PirDashboardWebMessages.INITIAL_SCAN_STATUS.messageName)
 
     override fun process(
         jsMessage: JsMessage,
         jsMessaging: JsMessaging,
         jsMessageCallback: JsMessageCallback?,
     ) {
-        logcat { "PIR-WEB: GetCurrentUserProfileMessageHandler: process $jsMessage" }
+        logcat { "PIR-WEB: InitialScanStatusMessageHandler: process $jsMessage" }
 
         jsMessaging.onResponse(
             JsCallbackData(
@@ -57,12 +57,19 @@ class PirWebGetCurrentUserProfileMessageHandler @Inject constructor() :
                         PirDashboardWebConstants.PARAM_VERSION,
                         PirDashboardWebConstants.SCRIPT_API_VERSION,
                     )
-                    // TODO: Replace with actual user profile data
+                    // TODO: Replace with actual data
+                    putOpt(PARAM_RESULTS_FOUND, JSONArray())
+                    putOpt(PARAM_SCAN_PROGRESS, JSONArray())
                 },
                 featureName = jsMessage.featureName,
                 method = jsMessage.method,
                 id = jsMessage.id ?: "",
             ),
         )
+    }
+
+    companion object {
+        const val PARAM_RESULTS_FOUND = "resultsFound"
+        const val PARAM_SCAN_PROGRESS = "scanProgress"
     }
 }
